Move string centric Util class from wcs.v2_0 to dap4 package to preserve build independence. Mpved WCS libs to resources/WCS/2.0/lib
