+++ src/opendap/dap4/Dataset.java
@@ -29,8 +29,6 @@ package opendap.dap4;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import opendap.wcs.v2_0.Util;
-
 import java.util.*;
 import javax.xml.bind.annotation.*;
